backend/pyproject.toml
[project]
name = "phishing-mail-backend"
version = "0.1.0"
description = "Local runnable MVP backend for phishing email identification and handling"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "celery[redis]>=5.4.0,<6.0.0",
  "cryptography>=43.0.0,<44.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "openai>=1.75.0,<2.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "python-multipart>=0.0.12,<1.0.0",
  "redis>=5.1.0,<6.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
